refactor: move extension install state into extensions crate
