+++ crates/ironclaw_product_adapter_registry/Cargo.toml
@@ -3,7 +3,7 @@ name = "ironclaw_product_adapter_registry"
 version = "0.1.0"
 edition = "2024"
 rust-version = "1.92"
-description = "ProductAdapter host-api projection and installation registry for IronClaw Reborn"
+description = "ProductAdapter host-api projection for IronClaw Reborn extension manifests"
 authors = ["NEAR AI <support@near.ai>"]
 license = "MIT OR Apache-2.0"
 homepage = "https://github.com/nearai/ironclaw"
@@ -11,16 +11,14 @@ repository = "https://github.com/nearai/ironclaw"
 publish = false
 
 [dependencies]
-async-trait = "0.1"
-chrono = { version = "0.4", features = ["serde"] }
 ironclaw_extensions = { path = "../ironclaw_extensions", version = "0.1.0" }
 ironclaw_host_api = { path = "../ironclaw_host_api", version = "0.1.0" }
 ironclaw_product_adapters = { path = "../ironclaw_product_adapters", version = "0.1.0" }
 serde = { version = "1", features = ["derive"] }
 thiserror = "2"
-tokio = { version = "1", features = ["sync"] }
 toml = "0.8"
 
 [dev-dependencies]
+chrono = { version = "0.4", features = ["serde"] }
 serde_json = "1"
 tokio = { version = "1", features = ["macros", "rt", "sync"] }
