+++ crates/ironclaw_product_adapter_registry/src/lib.rs
@@ -1,12 +1,8 @@
-//! Generic extension installation registry with ProductAdapter projection.
-//!
-//! There is one store, one installation type, one manifest type. The
-//! ProductAdapter-specific path is a projection on top:
+//! ProductAdapter host-api projection over generic extension installation state.
 //!
 //! ```text
-//! ExtensionInstallationStore (generic, async, persistence-backed)
-//!   manifests    → ExtensionManifestRecord   (any ExtensionManifestV2)
-//!   installations → ExtensionInstallation    (any extension, keyed by ExtensionId)
+//! ironclaw_extensions::ExtensionInstallationStore
+//!   manifests/installations for any extension
 //!
 //! list_enabled_product_adapter_entries(store)
 //!   → filter enabled installations whose manifest carries ironclaw.product_adapter/v1
@@ -17,23 +13,27 @@
 #![forbid(unsafe_code)]
 
 use std::collections::{BTreeSet, HashMap};
-use std::fmt;
 use std::sync::Arc;
 
-use async_trait::async_trait;
-use chrono::{DateTime, Utc};
 use ironclaw_extensions::{
     ExtensionManifestV2, HostApiContractRegistry, HostApiId, HostApiManifestContract,
     HostApiMultiplicity, HostApiRefV2, ManifestSectionPath, ManifestSource, ManifestV2Error,
 };
-use ironclaw_host_api::{ExtensionId, HostPortCatalog, SecretHandle};
+use ironclaw_host_api::{ExtensionId, HostPortCatalog};
 use ironclaw_product_adapters::{
     AuthRequirement, DeclaredEgressTarget, EgressCredentialHandle, ProductAdapterCapabilities,
-    ProductAdapterId, ProductCapabilityFlag, ProductSurfaceKind, RedactedString,
+    ProductAdapterId, ProductCapabilityFlag, ProductSurfaceKind,
 };
-use serde::{Deserialize, Deserializer, Serialize};
+use serde::Deserialize;
 use thiserror::Error;
-use tokio::sync::RwLock;
+
+pub use ironclaw_extensions::{
+    ExtensionActivationState, ExtensionCredentialBinding, ExtensionCredentialHandle,
+    ExtensionHealthMessage, ExtensionHealthSnapshot, ExtensionHealthStatus, ExtensionInstallation,
+    ExtensionInstallationError, ExtensionInstallationId, ExtensionInstallationStore,
+    ExtensionManifestRecord, ExtensionManifestRef, InMemoryExtensionInstallationStore,
+    ManifestHash,
+};
 
 // ---------------------------------------------------------------------------
 // Constants
@@ -42,615 +42,33 @@ use tokio::sync::RwLock;
 pub const PRODUCT_ADAPTER_HOST_API_ID: &str = "ironclaw.product_adapter/v1";
 pub const PRODUCT_ADAPTER_SECTION_PREFIX: &str = "product_adapter";
 
-// ---------------------------------------------------------------------------
-// Shared primitives
-// ---------------------------------------------------------------------------
-
-#[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord, Hash, Serialize)]
-#[serde(transparent)]
-pub struct ManifestHash(String);
-
-impl ManifestHash {
-    pub fn new(value: impl Into<String>) -> Result<Self, RegistryError> {
-        let value = value.into();
-        validate_nonempty_noncontrol("manifest_hash", &value)?;
-        Ok(Self(value))
-    }
-
-    pub fn as_str(&self) -> &str {
-        &self.0
-    }
-}
-
-impl<'de> Deserialize<'de> for ManifestHash {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        let value = String::deserialize(deserializer)?;
-        Self::new(value).map_err(serde::de::Error::custom)
-    }
-}
-
-// ---------------------------------------------------------------------------
-// ExtensionManifestRecord — single manifest type for any extension
-// ---------------------------------------------------------------------------
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct ExtensionManifestRecord {
-    raw_toml: String,
-    manifest: ExtensionManifestV2,
-    product_adapters: Vec<ProductAdapterHostApiSection>,
+pub fn parse_product_adapter_manifest_record(
+    raw_toml: impl Into<String>,
+    source: ManifestSource,
+    host_port_catalog: &HostPortCatalog,
     manifest_hash: Option<ManifestHash>,
+) -> Result<ExtensionManifestRecord, RegistryError> {
+    let contract = Arc::new(ProductAdapterHostApiContract::new()?);
+    let mut contracts = HostApiContractRegistry::new();
+    contracts.register(contract)?;
+    ExtensionManifestRecord::from_toml_with_contracts(
+        raw_toml,
+        source,
+        host_port_catalog,
+        manifest_hash,
+        &contracts,
+    )
+    .map_err(|error| match error {
+        ExtensionInstallationError::Manifest(error) => RegistryError::Manifest(error),
+        other => RegistryError::Installation(other),
+    })
 }
 
-impl ExtensionManifestRecord {
-    /// Parse a single Extension Manifest v2 TOML document. ProductAdapter
-    /// host-api sections are projected automatically when present.
-    pub fn from_toml(
-        raw_toml: impl Into<String>,
-        source: ManifestSource,
-        host_port_catalog: &HostPortCatalog,
-        manifest_hash: Option<ManifestHash>,
-    ) -> Result<Self, RegistryError> {
-        let contract = Arc::new(ProductAdapterHostApiContract::new()?);
-        let mut contracts = HostApiContractRegistry::new();
-        contracts.register(contract)?;
-        Self::from_toml_with_contracts(
-            raw_toml,
-            source,
-            host_port_catalog,
-            manifest_hash,
-            &contracts,
-        )
-    }
-
-    /// Parse with a caller-supplied contract registry (for additional host API types).
-    pub fn from_toml_with_contracts(
-        raw_toml: impl Into<String>,
-        source: ManifestSource,
-        host_port_catalog: &HostPortCatalog,
-        manifest_hash: Option<ManifestHash>,
-        contracts: &HostApiContractRegistry,
-    ) -> Result<Self, RegistryError> {
-        let raw_toml = raw_toml.into();
-        let manifest = ExtensionManifestV2::parse_with_host_api_contracts(
-            &raw_toml,
-            source,
-            host_port_catalog,
-            contracts,
-        )?;
-        let product_adapters = project_product_adapter_sections(&raw_toml, &manifest)?;
-        Ok(Self {
-            raw_toml,
-            manifest,
-            product_adapters,
-            manifest_hash,
-        })
-    }
-
-    pub fn manifest(&self) -> &ExtensionManifestV2 {
-        &self.manifest
-    }
-
-    pub fn raw_toml(&self) -> &str {
-        &self.raw_toml
-    }
-
-    pub fn extension_id(&self) -> &ExtensionId {
-        &self.manifest.id
-    }
-
-    pub fn product_adapters(&self) -> &[ProductAdapterHostApiSection] {
-        &self.product_adapters
-    }
-
-    pub fn manifest_hash(&self) -> Option<&ManifestHash> {
-        self.manifest_hash.as_ref()
-    }
+pub fn product_adapter_sections(
+    record: &ExtensionManifestRecord,
+) -> Result<Vec<ProductAdapterHostApiSection>, RegistryError> {
+    project_product_adapter_sections(record.raw_toml(), record.manifest())
 }
-
-// ---------------------------------------------------------------------------
-// Generic extension installation state
-// ---------------------------------------------------------------------------
-
-#[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord, Hash, Serialize)]
-#[serde(transparent)]
-pub struct ExtensionInstallationId(String);
-
-impl ExtensionInstallationId {
-    pub fn new(value: impl Into<String>) -> Result<Self, RegistryError> {
-        let value = value.into();
-        validate_nonempty_noncontrol("installation_id", &value)?;
-        Ok(Self(value))
-    }
-
-    pub fn as_str(&self) -> &str {
-        &self.0
-    }
-}
-
-impl fmt::Display for ExtensionInstallationId {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.write_str(&self.0)
-    }
-}
-
-impl<'de> Deserialize<'de> for ExtensionInstallationId {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        let value = String::deserialize(deserializer)?;
-        Self::new(value).map_err(serde::de::Error::custom)
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum ExtensionActivationState {
-    Installed,
-    Disabled,
-    Enabled,
-}
-
-/// Opaque credential binding: maps a manifest-declared handle name to a
-/// host secret handle. Never stores raw secret material.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct ExtensionCredentialBinding {
-    credential_handle: EgressCredentialHandle,
-    secret_handle: SecretHandle,
-}
-
-impl ExtensionCredentialBinding {
-    pub fn new(credential_handle: EgressCredentialHandle, secret_handle: SecretHandle) -> Self {
-        Self {
-            credential_handle,
-            secret_handle,
-        }
-    }
-
-    pub fn credential_handle(&self) -> &EgressCredentialHandle {
-        &self.credential_handle
-    }
-
-    pub fn secret_handle(&self) -> &SecretHandle {
-        &self.secret_handle
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct ExtensionManifestRef {
-    extension_id: ExtensionId,
-    manifest_hash: Option<ManifestHash>,
-}
-
-impl ExtensionManifestRef {
-    pub fn new(extension_id: ExtensionId, manifest_hash: Option<ManifestHash>) -> Self {
-        Self {
-            extension_id,
-            manifest_hash,
-        }
-    }
-
-    pub fn extension_id(&self) -> &ExtensionId {
-        &self.extension_id
-    }
-
-    pub fn manifest_hash(&self) -> Option<&ManifestHash> {
-        self.manifest_hash.as_ref()
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct ExtensionHealthSnapshot {
-    status: ExtensionHealthStatus,
-    message: Option<RedactedString>,
-    checked_at: DateTime<Utc>,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum ExtensionHealthStatus {
-    Healthy,
-    Degraded,
-    Unhealthy,
-}
-
-impl ExtensionHealthSnapshot {
-    pub fn healthy() -> Self {
-        Self {
-            status: ExtensionHealthStatus::Healthy,
-            message: None,
-            checked_at: Utc::now(),
-        }
-    }
-
-    pub fn new(
-        status: ExtensionHealthStatus,
-        message: Option<RedactedString>,
-        checked_at: DateTime<Utc>,
-    ) -> Self {
-        Self {
-            status,
-            message,
-            checked_at,
-        }
-    }
-
-    pub fn status(&self) -> ExtensionHealthStatus {
-        self.status
-    }
-
-    pub fn message(&self) -> Option<&RedactedString> {
-        self.message.as_ref()
-    }
-
-    pub fn checked_at(&self) -> DateTime<Utc> {
-        self.checked_at
-    }
-}
-
-/// One installed extension with activation state and opaque credential bindings.
-/// Works for any extension kind — ProductAdapter filtering is a projection.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize)]
-pub struct ExtensionInstallation {
-    installation_id: ExtensionInstallationId,
-    extension_id: ExtensionId,
-    activation_state: ExtensionActivationState,
-    manifest_ref: ExtensionManifestRef,
-    credential_bindings: Vec<ExtensionCredentialBinding>,
-    health: ExtensionHealthSnapshot,
-    updated_at: DateTime<Utc>,
-}
-
-impl ExtensionInstallation {
-    pub fn new(
-        installation_id: ExtensionInstallationId,
-        extension_id: ExtensionId,
-        activation_state: ExtensionActivationState,
-        manifest_ref: ExtensionManifestRef,
-        credential_bindings: Vec<ExtensionCredentialBinding>,
-        updated_at: DateTime<Utc>,
-    ) -> Result<Self, RegistryError> {
-        if manifest_ref.extension_id() != &extension_id {
-            return Err(RegistryError::ManifestExtensionMismatch {
-                extension_id,
-                manifest_extension_id: manifest_ref.extension_id().clone(),
-            });
-        }
-        validate_bindings_unique(&credential_bindings)?;
-        Ok(Self {
-            installation_id,
-            extension_id,
-            activation_state,
-            manifest_ref,
-            credential_bindings,
-            health: ExtensionHealthSnapshot::healthy(),
-            updated_at,
-        })
-    }
-
-    pub fn installation_id(&self) -> &ExtensionInstallationId {
-        &self.installation_id
-    }
-
-    pub fn extension_id(&self) -> &ExtensionId {
-        &self.extension_id
-    }
-
-    pub fn activation_state(&self) -> ExtensionActivationState {
-        self.activation_state
-    }
-
-    pub fn manifest_ref(&self) -> &ExtensionManifestRef {
-        &self.manifest_ref
-    }
-
-    pub fn credential_bindings(&self) -> &[ExtensionCredentialBinding] {
-        &self.credential_bindings
-    }
-
-    pub fn health(&self) -> &ExtensionHealthSnapshot {
-        &self.health
-    }
-
-    pub fn updated_at(&self) -> DateTime<Utc> {
-        self.updated_at
-    }
-
-    fn set_activation_state(&mut self, state: ExtensionActivationState) {
-        self.activation_state = state;
-        self.updated_at = Utc::now();
-    }
-
-    fn set_health(&mut self, health: ExtensionHealthSnapshot) {
-        self.health = health;
-        self.updated_at = Utc::now();
-    }
-}
-
-impl<'de> Deserialize<'de> for ExtensionInstallation {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        #[derive(Deserialize)]
-        #[serde(deny_unknown_fields)]
-        struct Wire {
-            installation_id: ExtensionInstallationId,
-            extension_id: ExtensionId,
-            activation_state: ExtensionActivationState,
-            manifest_ref: ExtensionManifestRef,
-            credential_bindings: Vec<ExtensionCredentialBinding>,
-            health: ExtensionHealthSnapshot,
-            updated_at: DateTime<Utc>,
-        }
-        let wire = Wire::deserialize(deserializer)?;
-        if wire.manifest_ref.extension_id() != &wire.extension_id {
-            return Err(serde::de::Error::custom(
-                RegistryError::ManifestExtensionMismatch {
-                    extension_id: wire.extension_id,
-                    manifest_extension_id: wire.manifest_ref.extension_id().clone(),
-                },
-            ));
-        }
-        validate_bindings_unique(&wire.credential_bindings).map_err(serde::de::Error::custom)?;
-        Ok(Self {
-            installation_id: wire.installation_id,
-            extension_id: wire.extension_id,
-            activation_state: wire.activation_state,
-            manifest_ref: wire.manifest_ref,
-            credential_bindings: wire.credential_bindings,
-            health: wire.health,
-            updated_at: wire.updated_at,
-        })
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Single generic store trait
-// ---------------------------------------------------------------------------
-
-#[async_trait]
-pub trait ExtensionInstallationStore: Send + Sync {
-    async fn list_manifests(&self) -> Result<Vec<ExtensionManifestRecord>, RegistryError>;
-
-    async fn get_manifest(
-        &self,
-        extension_id: &ExtensionId,
-    ) -> Result<Option<ExtensionManifestRecord>, RegistryError>;
-
-    async fn upsert_manifest(&self, manifest: ExtensionManifestRecord)
-    -> Result<(), RegistryError>;
-
-    async fn list_installations(&self) -> Result<Vec<ExtensionInstallation>, RegistryError>;
-
-    async fn list_enabled_installations(&self)
-    -> Result<Vec<ExtensionInstallation>, RegistryError>;
-
-    async fn get_installation(
-        &self,
-        installation_id: &ExtensionInstallationId,
-    ) -> Result<Option<ExtensionInstallation>, RegistryError>;
-
-    async fn upsert_installation(
-        &self,
-        installation: ExtensionInstallation,
-    ) -> Result<(), RegistryError>;
-
-    async fn set_activation_state(
-        &self,
-        installation_id: &ExtensionInstallationId,
-        state: ExtensionActivationState,
-    ) -> Result<(), RegistryError>;
-
-    /// Replace the stored health snapshot without enforcing status/message
-    /// consistency; callers may report status codes before a diagnostic exists.
-    async fn update_health(
-        &self,
-        installation_id: &ExtensionInstallationId,
-        health: ExtensionHealthSnapshot,
-    ) -> Result<(), RegistryError>;
-}
-
-#[async_trait]
-impl<T> ExtensionInstallationStore for Arc<T>
-where
-    T: ExtensionInstallationStore + ?Sized,
-{
-    async fn list_manifests(&self) -> Result<Vec<ExtensionManifestRecord>, RegistryError> {
-        (**self).list_manifests().await
-    }
-
-    async fn get_manifest(
-        &self,
-        extension_id: &ExtensionId,
-    ) -> Result<Option<ExtensionManifestRecord>, RegistryError> {
-        (**self).get_manifest(extension_id).await
-    }
-
-    async fn upsert_manifest(
-        &self,
-        manifest: ExtensionManifestRecord,
-    ) -> Result<(), RegistryError> {
-        (**self).upsert_manifest(manifest).await
-    }
-
-    async fn list_installations(&self) -> Result<Vec<ExtensionInstallation>, RegistryError> {
-        (**self).list_installations().await
-    }
-
-    async fn list_enabled_installations(
-        &self,
-    ) -> Result<Vec<ExtensionInstallation>, RegistryError> {
-        (**self).list_enabled_installations().await
-    }
-
-    async fn get_installation(
-        &self,
-        installation_id: &ExtensionInstallationId,
-    ) -> Result<Option<ExtensionInstallation>, RegistryError> {
-        (**self).get_installation(installation_id).await
-    }
-
-    async fn upsert_installation(
-        &self,
-        installation: ExtensionInstallation,
-    ) -> Result<(), RegistryError> {
-        (**self).upsert_installation(installation).await
-    }
-
-    async fn set_activation_state(
-        &self,
-        installation_id: &ExtensionInstallationId,
-        state: ExtensionActivationState,
-    ) -> Result<(), RegistryError> {
-        (**self).set_activation_state(installation_id, state).await
-    }
-
-    async fn update_health(
-        &self,
-        installation_id: &ExtensionInstallationId,
-        health: ExtensionHealthSnapshot,
-    ) -> Result<(), RegistryError> {
-        (**self).update_health(installation_id, health).await
-    }
-}
-
-#[derive(Debug, Default, Clone)]
-pub struct InMemoryExtensionInstallationStore {
-    inner: Arc<RwLock<InMemoryState>>,
-}
-
-#[derive(Debug, Default)]
-struct InMemoryState {
-    manifests: HashMap<ExtensionId, ExtensionManifestRecord>,
-    installations: HashMap<ExtensionInstallationId, ExtensionInstallation>,
-}
-
-#[async_trait]
-impl ExtensionInstallationStore for InMemoryExtensionInstallationStore {
-    async fn list_manifests(&self) -> Result<Vec<ExtensionManifestRecord>, RegistryError> {
-        let inner = self.inner.read().await;
-        let mut manifests: Vec<_> = inner.manifests.values().cloned().collect();
-        manifests.sort_by(|a, b| a.extension_id().cmp(b.extension_id()));
-        Ok(manifests)
-    }
-
-    async fn get_manifest(
-        &self,
-        extension_id: &ExtensionId,
-    ) -> Result<Option<ExtensionManifestRecord>, RegistryError> {
-        Ok(self.inner.read().await.manifests.get(extension_id).cloned())
-    }
-
-    async fn upsert_manifest(
-        &self,
-        manifest: ExtensionManifestRecord,
-    ) -> Result<(), RegistryError> {
-        let mut inner = self.inner.write().await;
-        for installation in inner.installations.values() {
-            if installation.extension_id() == manifest.extension_id() {
-                validate_installation_against_one_manifest(&manifest, installation)?;
-            }
-        }
-        inner
-            .manifests
-            .insert(manifest.extension_id().clone(), manifest);
-        Ok(())
-    }
-
-    async fn list_installations(&self) -> Result<Vec<ExtensionInstallation>, RegistryError> {
-        let inner = self.inner.read().await;
-        let mut installations: Vec<_> = inner.installations.values().cloned().collect();
-        installations.sort_by(|a, b| a.installation_id().cmp(b.installation_id()));
-        Ok(installations)
-    }
-
-    async fn list_enabled_installations(
-        &self,
-    ) -> Result<Vec<ExtensionInstallation>, RegistryError> {
-        Ok(self
-            .list_installations()
-            .await?
-            .into_iter()
-            .filter(|i| i.activation_state() == ExtensionActivationState::Enabled)
-            .collect())
-    }
-
-    async fn get_installation(
-        &self,
-        installation_id: &ExtensionInstallationId,
-    ) -> Result<Option<ExtensionInstallation>, RegistryError> {
-        Ok(self
-            .inner
-            .read()
-            .await
-            .installations
-            .get(installation_id)
-            .cloned())
-    }
-
-    async fn upsert_installation(
-        &self,
-        installation: ExtensionInstallation,
-    ) -> Result<(), RegistryError> {
-        validate_bindings_unique(installation.credential_bindings())?;
-        let mut inner = self.inner.write().await;
-        validate_installation_against_manifest(&inner.manifests, &installation)?;
-        inner
-            .installations
-            .insert(installation.installation_id().clone(), installation);
-        Ok(())
-    }
-
-    async fn set_activation_state(
-        &self,
-        installation_id: &ExtensionInstallationId,
-        state: ExtensionActivationState,
-    ) -> Result<(), RegistryError> {
-        let mut inner = self.inner.write().await;
-        let InMemoryState {
-            manifests,
-            installations,
-        } = &mut *inner;
-        let installation = installations.get_mut(installation_id).ok_or_else(|| {
-            RegistryError::InstallationNotFound {
-                installation_id: installation_id.clone(),
-            }
-        })?;
-        if installation.activation_state() == state {
-            return Ok(());
-        }
-        if state == ExtensionActivationState::Enabled {
-            validate_installation_against_manifest(manifests, installation)?;
-        }
-        installation.set_activation_state(state);
-        Ok(())
-    }
-
-    async fn update_health(
-        &self,
-        installation_id: &ExtensionInstallationId,
-        health: ExtensionHealthSnapshot,
-    ) -> Result<(), RegistryError> {
-        self.inner
-            .write()
-            .await
-            .installations
-            .get_mut(installation_id)
-            .ok_or_else(|| RegistryError::InstallationNotFound {
-                installation_id: installation_id.clone(),
-            })?
-            .set_health(health);
-        Ok(())
-    }
-}
-
-// ---------------------------------------------------------------------------
-// ProductAdapter projection
-// ---------------------------------------------------------------------------
 
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct ProductAdapterHostApiSection {
@@ -805,12 +223,12 @@ pub async fn list_enabled_product_adapter_entries(
             .ok_or_else(|| RegistryError::UnknownManifest {
                 extension_id: installation.extension_id().clone(),
             })?;
-        validate_installation_against_one_manifest(manifest, &installation)?;
-        let adapters = manifest.product_adapters();
+        let adapters = project_product_adapter_sections(manifest.raw_toml(), manifest.manifest())?;
+        validate_installation_against_one_manifest(manifest, &installation, &adapters)?;
         if adapters.is_empty() {
             continue;
         }
-        for adapter in adapters {
+        for adapter in &adapters {
             entries.push(ProductAdapterRuntimeEntry::new(
                 installation.clone(),
                 adapter.clone(),
@@ -883,6 +301,8 @@ impl HostApiManifestContract for ProductAdapterHostApiContract {
 #[derive(Debug, Error, PartialEq, Eq)]
 pub enum RegistryError {
     #[error(transparent)]
+    Installation(#[from] ExtensionInstallationError),
+    #[error(transparent)]
     Manifest(#[from] ManifestV2Error),
     #[error("invalid {field}: {reason}")]
     InvalidValue { field: &'static str, reason: String },
@@ -926,21 +346,10 @@ pub enum RegistryError {
 // Internal validation helpers
 // ---------------------------------------------------------------------------
 
-fn validate_installation_against_manifest(
-    manifests: &HashMap<ExtensionId, ExtensionManifestRecord>,
-    installation: &ExtensionInstallation,
-) -> Result<(), RegistryError> {
-    let manifest = manifests.get(installation.extension_id()).ok_or_else(|| {
-        RegistryError::UnknownManifest {
-            extension_id: installation.extension_id().clone(),
-        }
-    })?;
-    validate_installation_against_one_manifest(manifest, installation)
-}
-
 fn validate_installation_against_one_manifest(
     manifest: &ExtensionManifestRecord,
     installation: &ExtensionInstallation,
+    product_adapters: &[ProductAdapterHostApiSection],
 ) -> Result<(), RegistryError> {
     if manifest.extension_id() != installation.manifest_ref().extension_id() {
         return Err(RegistryError::ManifestExtensionMismatch {
@@ -969,47 +378,25 @@ fn validate_installation_against_one_manifest(
     // For non-product-adapter manifests there are no declared credentials, so
     // any binding would be invalid; but installations on non-PA manifests are
     // allowed (they just must not carry bindings).
-    let declared: BTreeSet<_> = manifest
-        .product_adapters()
+    let declared: BTreeSet<_> = product_adapters
         .iter()
-        .flat_map(|pa| pa.required_credentials().iter().cloned())
+        .flat_map(|pa| {
+            pa.required_credentials()
+                .iter()
+                .map(EgressCredentialHandle::as_str)
+        })
         .collect();
     for binding in installation.credential_bindings() {
-        if !declared.contains(binding.credential_handle()) {
+        if !declared.contains(binding.credential_handle().as_str()) {
             return Err(RegistryError::UndeclaredCredentialHandle {
-                handle: binding.credential_handle().clone(),
+                handle: EgressCredentialHandle::new(binding.credential_handle().as_str()).map_err(
+                    |reason| RegistryError::InvalidValue {
+                        field: "credential_handle",
+                        reason: reason.to_string(),
+                    },
+                )?,
             });
         }
-    }
-    Ok(())
-}
-
-fn validate_bindings_unique(
-    credential_bindings: &[ExtensionCredentialBinding],
-) -> Result<(), RegistryError> {
-    let mut seen = BTreeSet::new();
-    for binding in credential_bindings {
-        if !seen.insert(binding.credential_handle.clone()) {
-            return Err(RegistryError::DuplicateCredentialBinding {
-                handle: binding.credential_handle.clone(),
-            });
-        }
-    }
-    Ok(())
-}
-
-fn validate_nonempty_noncontrol(field: &'static str, value: &str) -> Result<(), RegistryError> {
-    if value.is_empty() {
-        return Err(RegistryError::InvalidValue {
-            field,
-            reason: "must not be empty".to_string(),
-        });
-    }
-    if value.chars().any(|c| c == '\0' || c.is_control()) {
-        return Err(RegistryError::InvalidValue {
-            field,
-            reason: "must not contain control characters".to_string(),
-        });
     }
     Ok(())
 }
